requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
google-auth==1.6.3
idna==2.8
pyasn1==0.4.7
pyasn1-modules==0.2.6
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.12.0
urllib3==1.25.6
